+++ test/integration/events/recurrence_test.rb
@@ -73,4 +73,49 @@ class Events::RecurrenceTest < ActionDispatch::IntegrationTest
     assert(db_events.all? { |event| event.advanced_accounting.code == '706300' })
     assert(db_events.all? { |event| event.advanced_accounting.analytical_section == '9A54C' })
   end
+
+  test 'create a simple recurrent event' do
+    name = 'Fablab party'
+    post '/api/events',
+         params: {
+           event: {
+             title: name,
+             event_image_attributes: {
+               attachment: fixture_file_upload('/files/event/Party.jpg')
+             },
+             description: 'Come party tonight at the fablab...',
+             start_date: 2.weeks.from_now.utc,
+             end_date: 2.weeks.from_now.utc,
+             all_day: false,
+             start_time: '18:00',
+             end_time: '23:29',
+             amount: 20,
+             category_id: 2,
+             recurrence: 'month',
+             recurrence_end_at: 3.months.from_now.utc + 2.weeks
+           }
+         },
+         headers: upload_headers
+
+    # Check response format & status
+    assert_equal 201, response.status, response.body
+    assert_equal Mime[:json], response.content_type
+
+    # Check the events were correctly created
+    db_events = Event.where(title: name)
+    assert_equal 4, db_events.count
+    assert(db_events.all? { |event| !event.event_image.attachment.nil? })
+    assert(db_events.all? { |event| !event.description.empty? })
+    assert(db_events.all? { |event| event.availability.start_at.to_date >= 2.weeks.from_now.to_date })
+    assert(db_events.all? { |event| event.availability.start_at.to_date <= 3.months.from_now.end_of_day.to_date + 2.weeks })
+    assert(db_events.all? { |event| event.availability.end_at.to_date >= 2.weeks.from_now.to_date })
+    assert(db_events.all? { |event| event.availability.end_at.to_date <= 3.months.from_now.end_of_day.to_date + 2.weeks })
+    assert(db_events.none?(&:all_day?))
+    assert(db_events.all? { |event| event.amount == 2000 })
+    assert(db_events.all? { |event| event.event_theme_ids.empty? })
+    assert(db_events.all? { |event| event.category_id == 2 })
+    assert(db_events.all? { |event| event.age_range_id.nil? })
+    assert(db_events.all? { |event| event.event_files.count.zero? })
+    assert(db_events.all? { |event| event.event_price_categories.count.zero? })
+  end
 end