+++ test/integration/plans/create_plan_test.rb
@@ -8,28 +8,26 @@ class CreatePlanTest < ActionDispatch::IntegrationTest
     login_as(admin, scope: :user)
   end
 
-  test 'create a plan' do
+  test 'create a transversal partner plan' do
     plans_count = Plan.count
 
     post '/api/plans',
          params: {
            plan: {
              base_name: 'Abonnement test',
-             type: 'Plan',
-             group_id: 1,
+             type: 'PartnerPlan',
+             group_id: 'all',
              plan_category_id: nil,
              interval: 'week',
              interval_count: 2,
              amount: 10,
              ui_weight: 0,
              is_rolling: true,
-             monthly_payment: false,
+             monthly_payment: true,
              description: 'lorem ipsum dolor sit amet',
-             partner_id: '',
+             partner_id: 6,
              plan_file_attributes: {
-               id: nil,
-               _destroy: nil,
-               attachment: nil
+               attachment: fixture_file_upload('/files/document.pdf')
              }
            }
          }.to_json,
@@ -39,9 +37,50 @@ class CreatePlanTest < ActionDispatch::IntegrationTest
     assert_equal 201, response.status, response.body
     assert_equal Mime[:json], response.content_type
 
+    # Check the created plans
+    res = json_response(response.body)
+    assert_equal 2, res[:plan_ids].count
+    assert_equal plans_count + 2, Plan.count
+
+    plans = Plan.where(name: 'Abonnement test')
+    assert(plans.all? { |plan| !plan.plan_file.attachment.nil? })
+    assert(plans.all? { |plan| plan.type == 'PartnerPlan' })
+    assert(plans.all? { |plan| plan.partner_id == 6 })
+    assert(plans.all?(&:is_rolling))
+  end
+
+  test 'create a simple plan' do
+    plans_count = Plan.count
+
+    post '/api/plans',
+         params: {
+           plan: {
+             base_name: 'Abonnement simple',
+             type: 'Plan',
+             group_id: 1,
+             plan_category_id: nil,
+             interval: 'month',
+             interval_count: 1,
+             amount: 40
+           }
+         }.to_json,
+         headers: default_headers
+
+    # Check response format & status
+    assert_equal 201, response.status, response.body
+    assert_equal Mime[:json], response.content_type
+
     # Check the created plan
-    plan = json_response(response.body)
-    assert_equal Plan.last.id, plan[:plan_ids][0]
+    res = json_response(response.body)
     assert_equal plans_count + 1, Plan.count
+
+    plan = Plan.find(res[:plan_ids][0])
+    assert_not_nil plan
+    assert_equal 'Abonnement simple', plan.base_name
+    assert_not plan.is_rolling
+    assert_equal 1, plan.group_id
+    assert_equal 'month', plan.interval
+    assert_equal 1, plan.interval_count
+    assert_equal 4000, plan.amount
   end
 end